Fixed up builds on Debian 7
There are a couple of problems fixed in this commit. Firstly, Boost
was not marked as required. CMake now throws an error if you are missing
any required Boost components. Secondly, libiberty (from binutils) is used,
however it was not built by the binutils step, so I have added that in.
Also, I have added an include directive for Boost.

Unfortunately, there is a small error in the libiberty configure script,
so a patch has been included to fix this. The patch came from:
https://gcc.gnu.org/ml/gcc-patches/2014-01/msg00213.html